User can add a poem (#1)
* Allow User to add a Poem

This commit creates the first core function of this application: allow a
user to register and subsequently create a poem that's saved to the
application's database.

Most of what's contained in this commit is straightforward. User auth is
implemented by [Clearance](https://github.com/thoughtbot/clearance) and
the logic for poems is pretty anemic at this point. The values for
`Poem#status` are preliminary at this point, but their purpose is to
track a poem though the submission process.

The biggest thing worth explaining here is the removal of Skylight from
the application's Gemfile, relative to what
[Suspenders](https://github.com/thoughtbot/suspenders) output. It was
interfering with tests running and the project doesn't currently need a
monitoring solution. If it does Skylight should be re-added and
configured.

* Feedback commit

This commit adds a handful of small changes based on PR feedback.
- Semantic use of `link_to` in home/index.html.erb required adding a
`button` class to _buttons.scss.
- Using responders for a failed update required making the `alert`
messages availalbe in its translation file.