Create `Submission` & `SubmissionPacket` models
This commit creates 2 new models, `Submission` and `SubmissionPacket`.
Submission is the second core model opposite poem in the app and the
`SubmissionPacket` is the join model between `Poem` and `Submission`.
`SubmissionPacket` contains a status property that can't be calculated,
so a `has_and_belongs_to_many` association wasn't feasible.

`Submission::STATUSES` is a list of potential submission states taken
directly from Submittable's site. Restricting to those types will ease
the integration with Submittable down the road.
 `Poem::STATUSES` are all states I think make sense for a poem lifecycle
so far.

Finally, this commit removes an option from the SimpleForm initialize
about invalid handling that proved redundant alongside the errors
partial that exists in the app.

*Lessons and takeaways*
- Use constants where they make sense instead of jumping to the step of
creating helper methods.
- When passing params through strong_params, if the param value is an
array, that must be indicated in the strong_params method
- SimpleForm `associations` will create associated objects based on ids
with no extra input