Security and Convention Fixes
This commit makes a few changes:
- Updates the forms and controllers so that we're no longer passing a
`user_id` through the form as a hidden field and are instead assigning
the user in the controller.
- Updates the model validations to use `validates <field>, presence:
true` instead of `validates_presence_of`, because the latter is old
style and not preferred.
- Adds a new home.en.yml file for translations on the homepage

* Lessons*
- Passing the `user_id` of the current_user as a hidden form input is a
security risk. Don't do that anymore. Instead, the user should be
assigned in the controller. There are two strategies for it in this
commit, and the Hash merge is probably the preferable one.
- Use `validates` with options instead of validation specific methods
(unless using a custom validation)
- When creating constants, be sure to freeze string literals and arrays
that contain them whenever applicable