Create new poem from submission form
This commit allows a user to create a new poem when creating a
submission. There were a few failed approaches before this commit.
 - I wasn't able to get `accepts_nested_attributes_for` to properly assign
the `current_user` to the newly created poem and strong params didn't
provide an easy way to access the nested attributes for the new poem.
- Extracting this to a service object didn't address that problem and
didn't seem to fit with the eventual roadmap item of adding poems
through AJAX.
- Eventually, I'd like to create this as a transaction, but there's a
pattern I'm going to want to use of a misconception I'll want to
correct, because in thinking about the ability to add multiple poems via
AJAX, I was thinking towards a switch in the controller based on what
the request contains. That said, an AJAX option that just builds and
returns poems on the submission will work the way I want.