`SubmissionForm` uses `accepts_nested_attributes_for`
This commit updates the `SubmissionsController` to use
`accepts_nested_attributes_for` instead of the previous workaround for
it.

Lessons
====
- Prior to this commit, I hadn't declared the `has_many_through`
relationship for `Submission` and `Poem`. A `has_many_through` that uses
`accepts_nested_attributes_for` two nested attributes statements: One on
the model that's accepting the other attributes and **one on the join
model**.
- Rails doesn't automatically assign the same `user_id` to all nested
items with an `accepts_nested_attributes_for`. The
`SubmissionController` here takes an extra step to ensure all poems in
the submission belong to `current_user`.
- `reject_if` is the method to use if every submission with
`accepts_nested_attributes_for` if the nested model is optional.