Rename app to **Imitable**
The app had was still referring to itself by its old name, "Submish".
This commit changes that. It also required a change to the underlying
local database for anyone who's running a local copy of this
application. To do that, I had to drop into psql and run `ALTER
DATABASE <current name> RENAME TO <new name>`.